% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounder.R
\name{rounder}
\alias{rounder}
\title{Round to the nearest integer}
\usage{
rounder(x)
}
\arguments{
\item{x}{A numeric element or vector to round to the nearest integer}
}
\value{
An integer element or vector
}
\description{
Function \code{\link[base]{round}} rounds - by default -  the number \code{1.5} to
integer \code{2} and the number \code{2.5} also to integer \code{2}. The reason for this
behaviour is the IEC 60559 standard where a 5 is expected to be rounded to the
even digit. This function provides an integer rounding alternative to
\code{\link[base]{round}}.
}
\examples{
library(magrittr)
# rounds weirdly
c(0.5, 1.5, 2.5, 3.5) \%>\% round()
# rounds correctly
c(0.5, 1.5, 2.5, 3.5) \%>\% rounder()
}
\author{
Gerko Vink \email{g.vink@uu.nl} and Hanne Oberman \email{h.i.oberman@uu.nl}
}
