% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundR-package.R
\docType{package}
\name{roundR-package}
\alias{roundR}
\alias{roundR-package}
\title{roundR: Round Numeric Values to the Nearest Integer}
\description{
By default, R rounds numeric values to even integers, following the IEC 60559 standard. This package offers alternative functionality to round to the closest integer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanneoberman/roundR}
  \item Report bugs at \url{https://github.com/hanneoberman/roundR/issues}
}

}
\author{
\strong{Maintainer}: Hanne Oberman \email{h.i.oberman@uu.nl}

Authors:
\itemize{
  \item Gerko Vink \email{g.vink@uu.nl}
}

}
\keyword{internal}
