% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundR-package.R
\docType{package}
\name{roundR-package}
\alias{roundR}
\alias{roundR-package}
\title{roundR: Round Numeric Values to the Nearest Integer}
\description{
In R the default choice for rounding follows the IEC 60559 standard where a 5 is expected to be rounded to the even digit. This package offers alternative functionality to round to the closest integer.
}
\keyword{internal}
